Move PrivateKeyBundle, PrivateKeyBytes into @iov/bns
