+++ packages/iov-bns/src/types.ts
@@ -1,19 +1,19 @@
 import * as Long from "long";
+import { As } from "type-tagger";
 
 import { BcpCoin, FullSignature, FungibleToken, Nonce, TokenTicker } from "@iov/bcp-types";
 import { Int53 } from "@iov/encoding";
-import {
-  Algorithm,
-  PrivateKeyBundle,
-  PrivateKeyBytes,
-  PublicKeyBundle,
-  PublicKeyBytes,
-  SignatureBytes,
-} from "@iov/tendermint-types";
+import { Algorithm, PublicKeyBundle, PublicKeyBytes, SignatureBytes } from "@iov/tendermint-types";
 
 import * as codecImpl from "./codecimpl";
 import { encodePubkey } from "./encode";
 import { InitData } from "./normalize";
+
+export type PrivateKeyBytes = Uint8Array & As<"private-key">;
+export interface PrivateKeyBundle {
+  readonly algo: Algorithm;
+  readonly data: PrivateKeyBytes;
+}
 
 export interface Result {
   readonly key: Uint8Array;
