+++ packages/iov-bns/types/encode.d.ts
@@ -1,6 +1,7 @@
 import { FungibleToken, SignedTransaction, UnsignedTransaction } from "@iov/bcp-types";
-import { PrivateKeyBundle, PublicKeyBundle } from "@iov/tendermint-types";
+import { PublicKeyBundle } from "@iov/tendermint-types";
 import * as codecImpl from "./codecimpl";
+import { PrivateKeyBundle } from "./types";
 export declare function encodePubkey(publicKey: PublicKeyBundle): codecImpl.crypto.IPublicKey;
 export declare function encodePrivkey(privateKey: PrivateKeyBundle): codecImpl.crypto.IPrivateKey;
 export declare function encodeToken(token: FungibleToken): codecImpl.x.Coin;
