+++ packages/iov-tendermint-types/src/keys.ts
@@ -5,12 +5,6 @@ export enum Algorithm {
   Secp256k1 = "secp256k1",
 }
 
-export type PrivateKeyBytes = Uint8Array & As<"private-key">;
-export interface PrivateKeyBundle {
-  readonly algo: Algorithm;
-  readonly data: PrivateKeyBytes;
-}
-
 export type PublicKeyBytes = Uint8Array & As<"public-key">;
 export interface PublicKeyBundle {
   readonly algo: Algorithm;
