+++ packages/iov-tendermint-types/types/keys.d.ts
@@ -3,11 +3,6 @@ export declare enum Algorithm {
     Ed25519 = "ed25519",
     Secp256k1 = "secp256k1"
 }
-export declare type PrivateKeyBytes = Uint8Array & As<"private-key">;
-export interface PrivateKeyBundle {
-    readonly algo: Algorithm;
-    readonly data: PrivateKeyBytes;
-}
 export declare type PublicKeyBytes = Uint8Array & As<"public-key">;
 export interface PublicKeyBundle {
     readonly algo: Algorithm;
